Syntax sugar for asset plugins
