+++ packages/assets/src/index.ts
@@ -1,7 +1,7 @@
 export * from './Assets';
+export * from './AssetExtension';
 export * from './utils';
 export * from './cache';
 export * from './loader';
 export * from './resolver';
 export * from './detections';
-
