+++ packages/spritesheet/src/index.ts
@@ -1,4 +1,2 @@
-export * from './loadSpritesheet';
 export * from './Spritesheet';
-export * from './cacheSpritesheet';
-export * from './resolveSpriteSheetUrl';
+export * from './spritesheetAsset';
